Use BindingTargets, Properties, Actions everywhere
Don’t expose ModelCoordinator to any controller. Use actions that receive a project ID as input and provide a non-mutable Property or a BindingTarget to the entity (such as a goal or time report) whose values are to be read or updated.

Get rid of Relay and use MutableProperty + de-optionalized binding target and skipNil() where necessary.

Speaking of which, use pairs composed of 1. a private MutableProperty (with an associated type that is optional) and a nil initial value, and 2. a public/internal BindingTarget (with a non optional as associated type when it makes sense) across the board (the board of controllers anyway).
(Having properties as intermediate relays allows for timing flexibility making the connection between inputs and outputs on one side and the contained controllers, or views, or things that are available only when the view is loaded, on the other.)

Remove redundant SectionIndex, substituting it with Section.

Get rid of DisposableBag. CompositeDisposable is available. And I don’t need it anymore, for the moment.