Put all direct accesses to toggl API into TogglAPIAccess
Simplify and reduce responsibilities of ModelCoordinator. If I continue down this road maybe one day I will come up with a proper name for this thing.

Now the bad news.
The “visible” reordering of the project list when a goal is added, modified or deleted is (temporarily) lost. Not only that, but the selected project is deselected each time a goal changes. This is to avoid the contrived mess that having one property for ProjectsByGoals plus two Signals to signal full refresh and “clued” refresh introduced. I am going to find a cleaner solution, I promise.