Simplify list VC by complicating everything else
Just kidding. GoalsStore did acquire a new responsibility: receiving a colletion of project IDs and emitting a stream updates, either partial (single move) or full (whole thing sent, refresh needed). The “thing” is a “new” (if not fully unfamiliar) struct called ProjectIDsByGoals.

Just the IDs and the values that become needed are passed around from the models to the controllers. No more sharing collections of all projects and all goals.

Sorting is also fully sent to and clearly enclosed in the model layer where it belongs.