Use method to set all actions together.
Ditch `doAfterViewIsLoaded` because it was standing on unreliable foundations. There was at least a case in which the `isViewLoaded` KVO notification in which it depended was not firing. Plus the `isViewLoaded` value changes to `true` before the system invokes `viewDidLoad` and thus before any of the children controllers were loaded. It was working probably because it was called actually after the view loaded.

Add documentation to ProjectsListViewController.