Use connectInputs() in GoalPeriodsPreferencesVC.
Introduce a perhaps naively simple implementation for enforcing a single connection of the inputs to any controller. Performance is not important since each execution should be a one-off for a given controller. Scope capturing is a requirement. And treating multiple executions as a programmer error  on debug builds is nice to have.