Sanitize connection of interfaces to controllers.
Use .flatten(.latest) and SerialDisposable to ensure that only the last interface is bound for bindings both down and upstream.

Call private connectInterface() from viewDidLoad to ensure both a single call during the controller’s lifetime and that the bindings happen after the children controllers are available.

Get rid of enforceOnce (which was a kludge that wouldn’t do what it promised) and of the complication of using a boolean wrapped in a MutableProperty to wait until the children controllers were available.